+++ tests/TypeFormatTestProvider.php
@@ -18,36 +18,6 @@ trait TypeFormatTestProvider
         ];
     }
 
-    public function intersectionTypesProvider()
-    {
-        return [
-            // Object
-            ['Exception&object', new \RuntimeException, false],
-            ['object&countable', new \ArrayIterator([]), false],
-            ['object&stdClass', new \stdClass, false],
-            ['object&SimpleXMLElement', new \SimpleXMLElement('<foo>bar</foo>'), false],
-            // // Array
-            // ['array&countable', []],
-            // ['array&iterable&countable', [1, 2, 3]],
-            // // Bool
-            // ['bool&true', true],
-            // ['bool&false', false],
-            // // Bool, but invalid
-            // ['bool&true', false, false],
-            // ['bool&false', true, false],
-            // // Scalar
-            // ['scalar&bool&true', true],
-            // ['scalar&numeric&int', 123],
-            // ['scalar&float', 123.4],
-            // ['scalar&string', 'string'],
-            // // Scalar, but invalid
-            // ['scalar&bool&false', true, false],
-            // ['scalar&float', 123, false],
-            // ['scalar&int', 123.4, false],
-            // ['scalar&array', 'string', false],
-        ];
-    }
-
     public function allowedSymbolProvider()
     {
         return [
