+++ tests/TypeTest.php
@@ -3,6 +3,7 @@
 namespace Realodix\Assert\Tests;
 
 use Realodix\Assert\Assert;
+use Realodix\Assert\Exception\InvalidTypeDeclarationFormatException;
 
 class TypeTest extends TestCase
 {
@@ -92,6 +93,35 @@ class TypeTest extends TestCase
     {
         Assert::type($type, $value);
         $this->addToAssertionCount(1);
+    }
+
+    /**
+     * @dataProvider intersectionTypesProvider
+     */
+    public function testIntersectionTypes($type, $value, $pass = true)
+    {
+        if (! $pass) {
+            $this->expectException(InvalidTypeDeclarationFormatException::class);
+            $this->expectExceptionMessage(
+                'Intersection Types only support class and interface names as intersection members.'
+            );
+        }
+
+        Assert::type($type, $value);
+        $this->addToAssertionCount(1);
+    }
+
+    /**
+     * Intersection Types is called "pure" Intersection Types because combining Union
+     * Types and Intersection Types in the same declaration is not allowed.
+     */
+    public function testPureIntersectionTypes()
+    {
+        $this->expectException(InvalidTypeDeclarationFormatException::class);
+        $this->expectExceptionMessage(
+            "Combining '|' and '&' in the same declaration is not allowed."
+        );
+        Assert::type('numeric&int|string', 1);
     }
 
     /**
