+++ tests/TypeTestProvider.php
@@ -10,9 +10,6 @@ trait TypeTestProvider
     public function arrayProvider()
     {
         return [
-            // [[RuleSetInterface::class, RuleSetInterface::class], new Realodix], // duplikat
-            [RuleSetInterface::class, new Realodix], // boleh
-            [[RuleSetInterface::class], new Realodix],
             ['ArrayAccess', new \ArrayObject],
             ['Traversable', new \ArrayObject],
             ['Traversable', new \ArrayIterator([])],
@@ -185,6 +182,15 @@ trait TypeTestProvider
             ['string', ''],
             ['string', ' '],
             ['string', '0'],
+        ];
+    }
+
+    public function intersectionTypesProvider()
+    {
+        return [
+            // [[RuleSetInterface::class, RuleSetInterface::class], new Realodix], // duplikat
+            [RuleSetInterface::class, new Realodix], // boleh
+            [[RuleSetInterface::class], new Realodix],
         ];
     }
